requirements.txt
Flask
Flask-CORS
tensorflow
Pillow
numpy
